Created a script that will check my gmail inbox for today.
The script asks for your password (after inputting your email from details.json) and then retrieves email from today. The module pyzmail is then used to get the sender names and email addresses to display to the user and to, optionally, write these values to a local .json file (where they can be used for further processing or pure admiration). Finally, we print the names and associated addresses in the console for the user to see, even if they don't save the values.